Get OS_info_t::build_number from registry instead of kernel_info() in Windows
Currently, infoware gets OS version from kernel version in Windows. But the OS version of
Windows can be different from the kernel version.

In Windows, the fourth number of version string which is matched to OS_info_t::build_number
is called UBR. We can get the number from registry key UBR(Win8.1 and above) or BuildLabEx
(Win8 and below). There is no need for major version to check whether Windows is above 8.1
or not; If there is not UBR, then fall back to BuildLabEx.

This commit implements getting UBR from the registry.